+++ tkviews/node/__init__.py
@@ -1,7 +1,5 @@
 '''Tkviews nodes'''
 
-from .canvas import CanvasNode, Rectangle, Text, Image, Arc, Bitmap, Line, Oval, Polygon, Window
-from .containers import Container, View, For, If
 from .layout import LayoutSetup, Row, Column
 from .styles import Style, StyleItem, StyleError
 from .ttk import TtkWidgetNode, TtkStyle